add crud services for tests

+++ library/src/main/java/by/aurorasoft/replicator/annotation/ReplicatedService.java
@@ -31,6 +31,6 @@ public @interface ReplicatedService {
 
         int partitionCount() default 1;
 
-        int replicationFactor() default 1;
+        short replicationFactor() default 1;
     }
 }
